Windows builds carry 14 uncleaned compiler warnings (#1432)
This is the follow-through of #1354, the last MSVC warning sweep: the
Windows legs have been compiling with warnings ever since it closed.
Harvesting the `libhttrack (x64, Release)` and `(Win32, Release)` logs
off master turns up 14 distinct sites: 3 that only x64 sees (a 64-bit
`size_t` narrowing into MSVC's `unsigned int` or `int`), 10 that only
Win32 sees (a 64-bit `LLint`/`__int64` narrowing into a 32-bit `size_t`,
or a signed/unsigned comparison), and 1 common to both. Running both
widths matters: the two lists overlap in a single entry, so either width
alone hides most of the sweep.

Thirteen are ours and all thirteen are fixed. Where the value simply had
the wrong type I changed the type rather than casting — `ampargs` in
`fil_normalized_ex` and the bogus-link loop index in `htsparse` become
`size_t`, and the emergency signal-handler writers in `httrack.c` and
`htsbacktrace.c` count in `unsigned int`, which is what MSVC's `write()`
takes and what their fixed 256-byte buffer and string literals fit in
anyway. Casts appear only where a guard on the same path already bounds
the value: `opt->maxlink` is cast under its own `> 0` test, and
`cache_rstr_addr`'s length is cast under the clamp to `[0, 32768]` two
lines above it, matching the cast the very next line already had.
`r.size` gets `(size_t)` at two call sites in `httpmirror` — it counts
the bytes held at `r.adr`, so it fits `size_t` by construction, and line
1649 of the same function already spelled it that way.

Two chained assignments are split apart, because the chain was what
produced the narrowing: `opt->maxtime = opt->maxsite = 0` in
`htsselftest.c` (an `int` fed from an `LLint`, landed in #1079) and
`gotquery = ampargs = 1` in `htslib.c`. The second sat inside a
braceless `if`, so splitting it needs braces — worth flagging since
dropping them would have made `ampargs = 1` unconditional.

The one I left is `src/minizip/mztools.c(194,29)` C4267. `src/minizip/`
is patched in place here rather than kept pristine, with each delta
recorded in an `.orig`/`.diff` sidecar pair under `EXTRA_DIST`, so this
is fixable -- it is just not worth it: the warning is harmless, the code
is upstream's, and touching it would mean regenerating that pair for no
gain.

No real bug fell out of this. The one candidate worth naming is the ICP
reply path in `proxytrack.c`, where a message length lands in an
`unsigned short`: the reader caps an ICP message at 16 KB per RFC2186
and NUL-terminates its buffer, so `strlen() + 1` cannot reach 65535 and
the field cannot truncate. The cast there is now explicit and carries
that reasoning.

Verified on Linux, since MSVC is not available locally: gcc and clang
both build clean at `-Wall -Wextra`, and a warning-set diff against an
`origin/master` baseline built with identical flags shows nothing new
and five sign-compare classes removed — gcc was seeing the same
signedness bugs MSVC reports as C4018. `make check` is green at 389
tests. The Windows legs on this PR are the real check.

+++ src/htsbacktrace.c
@@ -369,7 +369,7 @@ void hts_backtrace_altstack_release(void *stack) {
 }
 
 /* Why the report has no frames: a silent gap reads as a handler that died. */
-static void print_no_trace(int fd, const char *msg, size_t len) {
+static void print_no_trace(int fd, const char *msg, unsigned int len) {
   if (write(fd, msg, len) != len) { /* no ssize_t: this is built on MSVC too */
     /* sorry GCC */
   }
